消除告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@16567 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/spd/spdFMP.cpp
@@ -391,7 +391,7 @@ found:
       }
       else
       {
-         while ( _itr < _totalRead  && _expect < sizeof( SPD_MAGIC ) )
+         while ( _itr < _totalRead  && (UINT32)_expect < sizeof( SPD_MAGIC ) )
          {
             if ( SPD_MAGIC[_expect] == _readBuf[_itr] )
             {
